Raise error on invalid numbers in python backend
The python backend was incorrectly *not* failing when some invalid JSON
numbers were found. The JSON specification constrains which numbers are
accepted, with leading zeros not allowed, and fractions needing at
least one digit before and after the decimal point.

This commit prevents invalid JSON numbers from being successfully parsed
by the python backend. The new checks add <10% of performance overhead
(when compared locally using the big_int_object and big_decimal_object
benchmarks, both of which are purposefully numerically heavy), but there
might be ways to improve it. However the python backend should be always
the last resource, as its speed is between 20 and 30 times that of the
fastest backends. And in any case, correctness comes before performance
anyway.

Part of this issue (leading zeros) was originally reported in #40.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>