Detect @types.coroutine read() as async files
Most of python awaitables are native coroutines (async def functions),
but the definition of an awaitable includes also generator functions
decorated with the @types.coroutine function, named generated-based
coroutines.

This small difference has bitten us already in the past (see #32 for
example), and while doing other work I realized we hadn't seen the end
of it. In particular, the automatic detection of the file object given
to the main entry point functions of ijson (basic_parse, parse, kvitems
and items) didn't recognize file objects with a generator-based read
coroutine as asynchronous file objects, and therefore incorrectly routed
them to be parsed using the synchronous family of functions, which of
course fails.

This commit adds detection of file objects with generator-base read()
coroutines on the main entry point functions of all backends, correctly
routing them through the corresponding *_async function.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>