Revert non-concatenating parse implementation
The new parse implementation that avoids concatenating path elements to
form a prefix against which items/kvitems can compare doesn't work when
an element name contains a dot in them (e.g., selecting prefix "0.1" in
{"0.1": 1}). This is because on the one hand parse() creates a list of
element names (in this case, ["0.1"]), while on the case of the prefix
it is converted into a list by always splitting on the dot (in this
case, ["0", "1"]).

This reverts commits 833a088 and 3866612, which introduced the new
functionality and ported the kvitems and items routines to use it,
correspondingly.

This change addresses #41.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>