State lack of event interception in async functions
The even interception mechanism currently works with the synchronous
generators and not with the async generators. The documentation doesn't
state this though, and has thus led to user confusion, like that
expressed in #48.

This commit adds a small statement to the documentation to clarify that
event interception currently works only with the synchronous generator
functions.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>